enhance: adjust layout about switch player

+++ src/app/window/imp.rs
@@ -46,35 +46,44 @@ impl ObjectImpl for Window {
         self.headerbar.set_decoration_layout(Some("menu:close"));
         self.menubutton.set_icon_name("open-menu-symbolic");
 
-        let disconnect = MenuItem::new(Some("Disconnect"), Some("app.disconnect"));
         let hide_decoration = MenuItem::new(Some("Hide Decoration"), Some("win.hide-decoration"));
         let passthrough = MenuItem::new(Some("Switch Passthrough"), Some("win.switch-passthrough"));
 
         let popover = PopoverMenu::builder()
             .accessible_role(gtk::AccessibleRole::MenuItemRadio)
             .build();
-        self.menu.append_item(&disconnect);
-        self.menu.append_item(&passthrough);
         self.menu
             .append_submenu(Some("Switch players"), &self.submenu);
+        self.menu.append_item(&passthrough);
         self.menu.append_item(&hide_decoration);
         popover.set_menu_model(Some(&self.menu));
+
         let _submenu = self.submenu.downgrade();
-        popover.connect_visible_submenu_notify(move |_| {
+        popover.connect_visible_submenu_notify(move |sub| {
+            if Some("Switch players") != sub.visible_submenu().as_deref() {
+                return;
+            }
             let Some(menu) = _submenu.upgrade() else {
                 return;
             };
             menu.remove_all();
 
+            let section = gio::Menu::new();
             let players = list_avaliable_players();
+            if !players.is_empty() {
+                let disconnect = MenuItem::new(Some("Disconnect"), Some("app.disconnect"));
+                menu.append_item(&disconnect);
+            }
+
             for player in players {
                 let item = MenuItem::new(Some(player.identity()), None);
                 item.set_action_and_target_value(
                     Some("app.connect"),
                     Some(&ToVariant::to_variant(player.identity())),
                 );
-                menu.append_item(&item);
+                section.append_item(&item);
             }
+            menu.append_section(None, &section);
         });
 
         self.menubutton.set_popover(Some(&popover));
